requirements.txt
fastapi
uvicorn
passlib
bcrypt==4.0.1
python-jose[cryptography]
slowapi
pydantic>=2.0